Prototype Effect bundles
There's no escape from builds after all. To have hot-reloading, all code
needs to be compiled and read. There's no way around it. v8 and JSC
hot-reloading mechanism are unreliable.

The idea is to watch a module that exports Effect<HttpApp>, compile it
on change, and replace the references in a top-level effect.

Because we're using this, we have a high safety all resources will be
properly managed between loads.

This change brings me closer to figuring out an optimal API for
arbitrary bundles, both for backend and frontend. I will have to expose
stream of changes for development to properly update top-level effect in
development. In production it will probably not be necessary as
bundles will be stored in the file system.

Since top-level Server will relay on two Effect bundles, one for
frontend one for backend, I will need to merge the routes and
reconstruct it each time one of the bundle changes.