Route: infer bindings around use()
We had to add RouteBindings symbol in order to enable proper inference
in RouteMount.Builder which extends RouteSet.

Phantom generics inference work when we infer directly on a type that
holds generics. When we try to infer a child, TS weirds out (in this case
it was resulting in `unknown`). However, for structural inference it
works so we now have RouteBindings phantom symbol.

In the future, we may use it to store static bindings at runtime to
optimize construction of routers or storing annotations
like schema.

I wonder if we can consolidate use() with filter()...